added in zeph's scripts but edited out the Recast specific bits
Credit for all of this goes to ZephyrWindSpirit, who smushed together most of the scripts in OG and Recast. Unfortunately since everything was in existing files, it would have undone progress made on Emmu's end. So I went through and merged everything in a way that adds in Zeph's stuff without removing Emmu's.

I also edited all the bits of the scripts which referred to mods we're not adding yet (Let's Do, Malum, etc). So this should work just fine in OG as it stands.